change mutable_batch with ref && improve point to row_data
